feat (car) : Add Cars.findWinner
